requirements.txt
python-dotenv~=1.1.1
azure-cognitiveservices-speech==1.32.1
pydantic==2.5.2
numpy==1.24.3
soundfile==0.12.1
python-slugify==8.0.1
